+++ app/src/main/java/com/yapp/bol/app/di/DataSourceModule.kt
@@ -2,14 +2,14 @@ package com.yapp.bol.app.di
 
 import com.yapp.bol.data.datasource.impl.RemoteDataSource
 import com.yapp.bol.data.datasource.auth.AuthDataSource
-import com.yapp.bol.data.datasource.auth.impl.AuthDataSourceImpl
+import com.yapp.bol.data.datasource.auth.AuthDataSourceImpl
 import com.yapp.bol.data.datasource.group.GroupDataSource
-import com.yapp.bol.data.datasource.group.impl.GroupDataSourceImpl
+import com.yapp.bol.data.datasource.group.GroupDataSourceImpl
 import com.yapp.bol.data.datasource.impl.RemoteDataSourceImpl
 import com.yapp.bol.data.datasource.rank.RankDataSource
-import com.yapp.bol.data.datasource.rank.impl.RankDataSourceImpl
+import com.yapp.bol.data.datasource.rank.RankDataSourceImpl
 import com.yapp.bol.data.datasource.setting.SettingDataSource
-import com.yapp.bol.data.datasource.setting.impl.SettingDataSourceImpl
+import com.yapp.bol.data.datasource.setting.SettingDataSourceImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
