+++ app/src/main/java/com/yapp/bol/app/di/RepositoryModule.kt
@@ -1,7 +1,7 @@
 package com.yapp.bol.app.di
 
 import com.yapp.bol.data.repository.RepositoryImpl
-import com.yapp.bol.data.repository.auth.AuthRepositoryImpl
+import com.yapp.bol.data.repository.AuthRepositoryImpl
 import com.yapp.bol.domain.repository.AuthRepository
 import com.yapp.bol.domain.repository.Repository
 import com.yapp.bol.data.repository.GroupRepositoryImpl
