+++ data/src/main/java/com/yapp/bol/data/datasource/auth/AuthDataSourceImpl.kt
@@ -1,12 +1,16 @@
-package com.yapp.bol.data.datasource.auth.impl
+package com.yapp.bol.data.datasource.auth
 
-import com.yapp.bol.data.datasource.auth.AuthDataSource
-import com.yapp.bol.data.datasource.auth.impl.datastore.AuthManager
+import com.yapp.bol.data.datasource.auth.LoginType.toDomain
+import com.yapp.bol.data.model.login.LoginRequest
+import com.yapp.bol.data.model.login.LoginResponse
+import com.yapp.bol.data.remote.AuthApi
+import com.yapp.bol.data.remote.MatchApi
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class AuthDataSourceImpl @Inject constructor(
     private val authManager: AuthManager,
+    private val authApi: AuthApi,
 ) : AuthDataSource {
 
     override val accessToken: Flow<String> = authManager.accessKey
@@ -27,5 +31,9 @@ class AuthDataSourceImpl @Inject constructor(
 
     override suspend fun deleteRefreshToken() {
         authManager.deleteRefreshToken()
+    }
+
+    override suspend fun login(type: String, token: String): LoginResponse? {
+        return authApi.postOAuthApi(LoginRequest(type.toDomain(), token)).body()
     }
 }
