+++ data/src/main/java/com/yapp/bol/data/datasource/match/MatchDataSource.kt
@@ -1,0 +1,8 @@
+package com.yapp.bol.data.datasource.match
+
+import com.yapp.bol.data.model.match.MatchApiRequest
+
+interface MatchDataSource {
+
+    suspend fun postMatch(matchApiRequest: MatchApiRequest)
+}
