+++ data/src/main/java/com/yapp/bol/data/remote/GameApi.kt
@@ -1,0 +1,14 @@
+package com.yapp.bol.data.remote
+
+import com.yapp.bol.data.model.group.response.GameApiResponse
+import retrofit2.Response
+import retrofit2.http.GET
+import retrofit2.http.Path
+
+interface GameApi {
+
+    @GET("/v1/group/{groupId}/game")
+    suspend fun getGameList(
+        @Path("groupId") groupId: Int,
+    ): Response<GameApiResponse>
+}
